organize for more consistent API
