+++ model/interface.go
@@ -2,13 +2,15 @@ package model
 
 import "io"
 
+type Stmt interface {
+	ID() string
+	WriteTo(io.Writer) (int64, error)
+}
+type Stmts []Stmt
+
 type maybeString struct {
 	Valid bool
 	Value string
-}
-
-type WriteTo interface {
-	WriteTo(io.Writer) (int64, error)
 }
 
 type ColumnContainer interface {
@@ -17,8 +19,8 @@ type ColumnContainer interface {
 }
 
 type Index interface {
+	Stmt
 	ColumnContainer
-	WriteTo
 
 	HasName() bool
 	HasSymbol() bool
@@ -112,3 +114,135 @@ const (
 	ReferenceOptionSetNull
 	ReferenceOptionNoAction
 )
+
+type Table interface {
+	Stmt
+
+	Name() string
+	IsTemporary() bool
+	SetTemporary(bool)
+	IsIfNotExists() bool
+	SetIfNotExists(bool)
+
+	AddColumn(TableColumn)
+	Columns() chan TableColumn
+	AddIndex(Index)
+	Indexes() chan Index
+	AddOption(TableOption)
+	Options() chan TableOption
+
+	LookupColumn(string) (TableColumn, bool)
+	LookupIndex(string) (Index, bool)
+}
+
+type TableOption interface {
+	Stmt
+	Key() string
+	Value() string
+}
+
+type table struct {
+	name        string
+	temporary   bool
+	ifnotexists bool
+	columns     []TableColumn
+	indexes     []Index
+	options     []TableOption
+}
+
+type tableopt struct {
+	key   string
+	value string
+}
+
+type NullState int
+
+const (
+	NullStateNone NullState = iota
+	NullStateNull
+	NullStateNotNull
+)
+
+type Length interface {
+	HasDecimal() bool
+	Decimal() string
+	SetDecimal(string)
+	Length() string
+}
+
+type length struct {
+	decimals maybeString
+	length   string
+}
+
+type TableColumn interface {
+	Stmt
+
+	Name() string
+	Type() ColumnType
+	SetType(ColumnType)
+
+	HasLength() bool
+	Length() Length
+	SetLength(Length)
+	HasCharacterSet() bool
+	CharacterSet() string
+	HasCollation() bool
+	Collation() string
+	HasDefault() bool
+	Default() string
+	SetDefault(string)
+	HasComment() bool
+	Comment() string
+	SetComment(string)
+
+	NullState() NullState
+	SetNullState(NullState)
+
+	IsAutoIncrement() bool
+	SetAutoIncrement(bool)
+	IsBinary() bool
+	SetBinary(bool)
+	IsKey() bool
+	SetKey(bool)
+	IsPrimary() bool
+	SetPrimary(bool)
+	IsUnique() bool
+	SetUnique(bool)
+	IsUnsigned() bool
+	SetUnsigned(bool)
+	IsZeroFill() bool
+	SetZeroFill(bool)
+
+}
+
+type tablecol struct {
+	name         string
+	typ          ColumnType
+	length       Length
+	nullstate    NullState
+	charset      maybeString
+	collation    maybeString
+	defaultValue maybeString
+	comment      maybeString
+	autoincr     bool
+	binary       bool
+	key          bool
+	primary      bool
+	unique       bool
+	unsigned     bool
+	zerofill     bool
+}
+
+type Database interface {
+	Stmt
+
+	Name() string
+	IsIfNotExists() bool
+	SetIfNotExists(bool)
+}
+
+type database struct {
+	name        string
+	ifnotexists bool
+}
